Fixes parsing of uneven-byte length invoices
The old byte string reader has a bug when reading data whose length is
not a factor of 5.

For example, reading 7 bits, given two 5 bit bytes:
11111 00000

The existing implementation performs the following:
1. Read 7 bits: 1111100
2. Skip final padding, so the final result is only 7 bits
3. Apply `Integer.parseInt("1111100")`, which _does not pad_, so the
   final result is off by 1 bit.

This results in an incorrect hexadecimal value of 7C.

The new version in this patch accounts for this missing bit by padding
the return value:
1. Read 7 bits: 1111100
2. Pad: 11111000
3. Apply Integer.parseInt("11111000"), so the value is no longer off
   by 1 bit.

This results in a correct hexadecimal value of F8.

This bug results in:
- Failure to retrieve the correct signature when parsing an invoice
- Failure to derive the correct public key from the signature, because
  the signature bits are incorrect

Changes:
- Fixes failure to pad end of byte string when there are an uneven
  number of bits to process
- Signatures now include the ECDSA recovery ID
- By fixing padding/bit misalignment bugs, the public key can be
  correctly derived from the signature